+++ NEWS.md
@@ -23,6 +23,7 @@ New features
         UUR, RUR, RDR, DDR, DDL, LDL, LUL, UUL are diagonal moves for a hexagonal board
 
   - ``N?PieceSpec`` is now a shortcut for ``N&?PieceSpec``
+  - An empty PieceSpec now refers to the "last active" piece(s)
 
 * New function ``view_game()`` provides a simple command-line PPN viewer/editor.
 * New function ``cat_game()`` renders a plaintext animation of a game within the terminal.